Split email tests into focused files and update guide to document ticket attachments
Split the monolithic email.test.ts (733 lines) into four focused test files
in test/lib/email/ with shared helpers:
- send-email.test.ts: provider-specific request formatting and error handling
- attachments.test.ts: SVG ticket data building and attachment generation
- config.test.ts: email config resolution from DB settings and env vars
- registration.test.ts: end-to-end registration email flow

Removed tautological isEmailProvider tests that only verified a Set contains
its known values. Added three meaningful tests: purchaseOnly flag propagation,
currency passthrough, and error resilience (Promise.allSettled ensures one
failed email doesn't block others).

Updated the admin guide's "What does the confirmation email contain?" section
to document SVG ticket attachments and purchase-only QR code omission, which
were present in the code but missing from the guide.
